refaktoring, fix bugov, pridany mutex a print statements, uprava validacii
